Set up inferno shell profile
/lib/sh/profile is similar to /etc/profile in that it is run
to set up shells created in inferno. This one now does the
job of setting the user home directory and running the user's
local profile if they have one.

Since xu401 is the root node of the cluster, it's been added
to the ndb as the default signer. Note, for this to work it
has to be added to /etc/hosts in the underlying os for the dns
resolution