Tweak C++ build flags for arm
Adds a number of build flags specific for the armv7 chips on
the xu4's, as well as some debug flags and an optimization
flag. The debug flags were added in an attempt to determine
what was going wrong in JIT compilation, but that appears
to be too difficult of a problem for now, so if we're going
to run interpreted, it might as well be faster.