Make bind-mounted directories writable
Working with multiple users is making things difficult. Currently the
process is to build a docker container with user inferno, and then bind
mount /lib, /usr, and /keydb into the container when running inferno.
The issue arises when the uid of the inferno user (1000) does not match
the uid of the user who cloned the repo; this results in the inferno os
not being able to write to those directories. The dockerfile contains
two build args (GID, UID) to alleviate this issue, but when the user
building the image is root (0, 0), it's not possible to propogate those
values into the inferno user in the container. A better solution would
be to not run inferno as a root user, but for the sake of progress and
ease of use, we're just going to make these bound directories writable
by anyone so that the inferno os can use them.