Split inferno user into 'head' and 'worker' user
Our goal is to create a cluster of odroids with one being the head
node and the rest being worker nodes. While some of the workers will
be treated as file servers and others as cpus, we can in reality
configure them all the same, and then let the head node mount each
as appropriate. But for this to happen automatically, the head node
will need a more complex configuration than the worker nodes.

Since the current usr/inferno setup is similar to what we want the
workers to be, it was renamed to usr/worker, and 'worker' was set
to be the default user name when building the docker image.